refactor: move tx mapping to a service
