+++ packages/snap/package.json
@@ -53,7 +53,7 @@
     "@solana-program/compute-budget": "0.7.0",
     "@solana-program/system": "0.7.0",
     "@solana-program/token": "0.5.1",
-    "@solana-program/token-2022": "^0.4.2",
+    "@solana-program/token-2022": "^0.5.0",
     "@solana/kit": "2.1.0",
     "@testing-library/jest-dom": "^6.6.2",
     "@types/express": "^5.0.0",
