+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-solana-wallet.git"
   },
   "source": {
-    "shasum": "d+6vu8JZFXJYlb/O7ccnFtJoS+jbfw/6xhTnBtsfbBY=",
+    "shasum": "9BnNcD31FNaezWvIyswLc7rxEf7rty255cJu0eD2kAY=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
