+++ packages/snap/src/core/services/assets/TokenHelper.ts
@@ -7,7 +7,16 @@ import type { Account, Address, Lamports } from '@solana/kit';
 import { lamports, unwrapOption } from '@solana/kit';
 import BigNumber from 'bignumber.js';
 
+import type { Network } from '../../constants/solana';
+import type { SolanaConnection } from '../connection';
+
 export class TokenHelper {
+  readonly #connection: SolanaConnection;
+
+  constructor(connection: SolanaConnection) {
+    this.#connection = connection;
+  }
+
   /**
    * Converts a UI amount to the raw amount in lamports for a given mint, accounting for
    * any special mint extensions such as interest bearing or scaled UI amount mints.
@@ -19,11 +28,45 @@ export class TokenHelper {
    *
    * Based on the implementation from @solana-labs token-2022.
    * @see https://github.com/solana-program/token-2022/blob/rust-legacy%40v0.17.0/clients/js/src/amountToUiAmount.ts#L329
-   * @param mintAccount - The mint account containing extension and decimal information.
+   * @param mintAddress - The mint address.
+   * @param network - The network to fetch the mint account from.
    * @param uiAmount - The UI amount to convert.
    * @returns The amount in lamports as a Lamports type.
    */
-  static uiAmountToAmountForMint(
+  async uiAmountToAmountForMint(
+    mintAddress: string,
+    network: Network,
+    uiAmount: string,
+  ): Promise<Lamports> {
+    const mintAccount = await this.#connection.fetchMint(mintAddress, network);
+    return this.#uiAmountToAmountForMintAccount(mintAccount, uiAmount);
+  }
+
+  /**
+   * Reverse operation of {@link uiAmountToAmountForMint}.
+   * Instead of re-implementing the logic, we use a mathematical trick to find the multiplier.
+   * @param mintAddress - The mint address.
+   * @param network - The network to fetch the mint account from.
+   * @param amount - The amount in lamports to convert.
+   * @returns The UI amount as a string.
+   */
+  async amountToUiAmountForMint(
+    mintAddress: string,
+    network: Network,
+    amount: Lamports,
+  ): Promise<string> {
+    const mintAccount = await this.#connection.fetchMint(mintAddress, network);
+    return this.#amountToUiAmountForMintAccount(mintAccount, amount);
+  }
+
+  /**
+   * Internal method that works with a mint account directly.
+   * Converts a UI amount to the raw amount in lamports for a given mint account.
+   * @param mintAccount - The mint account.
+   * @param uiAmount - The UI amount.
+   * @returns The amount in lamports as a Lamports type.
+   */
+  #uiAmountToAmountForMintAccount(
     mintAccount: Account<Mint, Address>,
     uiAmount: string,
   ): Lamports {
@@ -94,7 +137,7 @@ export class TokenHelper {
       // This should never happen due to the conditions above
       throw new Error('Unknown mint extension type');
     } catch (error: any) {
-      const decimals = mintAccount.data?.decimals ?? 9;
+      const decimals = mintAccount?.data?.decimals ?? 9;
       return lamports(
         BigInt(
           BigNumber(uiAmount)
@@ -106,13 +149,13 @@ export class TokenHelper {
   }
 
   /**
-   * Reverse operation of {@link uiAmountToAmountForMint}.
-   * Instead of re-impleting the logic, we use a mathematical trick to find the multiplier.
-   * @param mintAccount - The mint account containing extension and decimal information.
-   * @param amount - The amount in lamports to convert.
+   * Internal method that works with a mint account directly.
+   * Converts an amount in lamports to a UI amount for a given mint account.
+   * @param mintAccount - The mint account.
+   * @param amount - The amount in lamports.
    * @returns The UI amount as a string.
    */
-  static amountToUiAmountForMint(
+  #amountToUiAmountForMintAccount(
     mintAccount: Account<Mint, Address>,
     amount: Lamports,
   ): string {
@@ -120,7 +163,7 @@ export class TokenHelper {
       const fakeUiAmount = 1000000000n;
 
       const multiplier = BigNumber(
-        TokenHelper.uiAmountToAmountForMint(
+        this.#uiAmountToAmountForMintAccount(
           mintAccount,
           fakeUiAmount.toString(),
         ).toString(),
