+++ packages/snap/src/core/services/assets/index.ts
@@ -1,2 +1,3 @@
 export * from './AssetsRepository';
 export * from './AssetsService';
+export * from './TokenHelper';
