+++ packages/snap/src/core/services/transactions/index.ts
@@ -1,2 +1,3 @@
+export * from './TransactionMapper';
 export * from './TransactionsRepository';
 export * from './TransactionsService';
