+++ packages/snap/src/core/test/mocks/asset-entities.ts
@@ -1,10 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import type { FungibleAssetMetadata } from '@metamask/snaps-sdk';
-import type { CaipAssetType } from '@metamask/utils';
-import { TOKEN_PROGRAM_ADDRESS, type Mint } from '@solana-program/token';
-import { TOKEN_2022_PROGRAM_ADDRESS } from '@solana-program/token-2022';
-import type { Account } from '@solana/kit';
-import { address, lamports } from '@solana/kit';
 
 import type { AssetEntity, NativeAsset, TokenAsset } from '../../../entities';
 import { KnownCaip19Id, Network } from '../../constants/solana';
@@ -96,44 +91,3 @@ export const SOLANA_MOCK_TOKEN_METADATA: Record<string, FungibleAssetMetadata> =
       ],
     },
   };
-
-// Sample responses from fetchMint method, exported by @solana-program/token
-export const MOCK_FETCH_MINT_RESPONSES: Record<CaipAssetType, Account<Mint>> = {
-  [KnownCaip19Id.UsdcLocalnet]: {
-    executable: false,
-    lamports: lamports(390030797122n),
-    programAddress: TOKEN_PROGRAM_ADDRESS,
-    space: 82n,
-    address: address('4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU'), // Localnet USDC
-    data: {
-      mintAuthority: {
-        __option: 'Some',
-        value: address('BJE5MMbqXjVwjAF7oxwPYXnTXDyspzZyt4vwenNw5ruG'),
-      },
-      supply: 9011200678169477n,
-      decimals: 6,
-      isInitialized: true,
-      freezeAuthority: {
-        __option: 'Some',
-        value: address('7dGbd2QZcCKcTndnHcTL8q7SMVXAkp688NTQYwrRCrar'),
-      },
-    },
-  },
-  [KnownCaip19Id.Ai16zMainnet]: {
-    executable: false,
-    lamports: lamports(18364570235n),
-    programAddress: TOKEN_2022_PROGRAM_ADDRESS,
-    space: 408n,
-    address: address('HeLp6NuQkmYB4pYWo2zYs22mESHXPQYzXbB8n4V98jwC'), // Localnet ai16z
-    data: {
-      mintAuthority: {
-        __option: 'Some',
-        value: address('AZtt8LUScEAG74iKnPNRuYgQhwmGJhAf6yUkAXjAd8sp'),
-      },
-      supply: 1099998595821532811n,
-      decimals: 9,
-      isInitialized: true,
-      freezeAuthority: { __option: 'None' },
-    },
-  },
-} as const;
