+++ packages/snap/src/features/send/transactions/SendSplTokenBuilder.ts
@@ -1,4 +1,4 @@
-import { assert, Duration } from '@metamask/utils';
+import { assert } from '@metamask/utils';
 import {
   getSetComputeUnitLimitInstruction,
   getSetComputeUnitPriceInstruction,
@@ -8,33 +8,19 @@ import {
   getCreateAssociatedTokenIdempotentInstruction,
   getTransferCheckedInstruction,
 } from '@solana-program/token';
-import type { Mint } from '@solana-program/token-2022';
-import { fetchMint } from '@solana-program/token-2022';
-import type {
-  CompilableTransactionMessage,
-  GetAccountInfoApi,
-  GetTokenAccountsByOwnerApi,
-  Rpc,
-} from '@solana/kit';
+import type { CompilableTransactionMessage } from '@solana/kit';
 import {
   appendTransactionMessageInstructions,
   createKeyPairSignerFromPrivateKeyBytes,
   createTransactionMessage,
-  fetchJsonParsedAccount,
   pipe,
   prependTransactionMessageInstructions,
   setTransactionMessageFeePayer,
   setTransactionMessageLifetimeUsingBlockhash,
-  type Account,
   type Address,
-  type MaybeAccount,
-  type MaybeEncodedAccount,
 } from '@solana/kit';
 
-import type { ICache } from '../../../core/caching/ICache';
-import { useCache } from '../../../core/caching/useCache';
-import type { Serializable } from '../../../core/serialization/types';
-import { TokenHelper } from '../../../core/services/assets/TokenHelper';
+import type { TokenHelper } from '../../../core/services';
 import type { SolanaConnection } from '../../../core/services/connection';
 import type { TransactionHelper } from '../../../core/services/execution/TransactionHelper';
 import { deriveSolanaKeypair } from '../../../core/utils/deriveSolanaKeypair';
@@ -45,13 +31,13 @@ import type {
 } from './ISendTransactionBuilder';
 
 export class SendSplTokenBuilder implements ISendTransactionBuilder {
+  readonly #tokenHelper: TokenHelper;
+
   readonly #connection: SolanaConnection;
 
   readonly #transactionHelper: TransactionHelper;
 
   readonly #logger: ILogger;
-
-  readonly #cache: ICache<Serializable>;
 
   /**
    * The transaction built here consumes up to ~30,000 compute units when just transferring
@@ -62,25 +48,16 @@ export class SendSplTokenBuilder implements ISendTransactionBuilder {
 
   readonly #computeUnitPriceMicroLamportsPerComputeUnit = 10000n;
 
-  /**
-   * We cache the mint a token accounts for 1 minute to avoid making too many requests to the RPC.
-   * It's needed because in the Send flow, a message is rebuilt every time the user inputs an amount.
-   */
-  readonly #cacheTtlsMilliseconds = {
-    mintAccount: Duration.Minute,
-    tokenAccount: Duration.Minute,
-  };
-
   constructor(
+    tokenHelper: TokenHelper,
     connection: SolanaConnection,
     transactionHelper: TransactionHelper,
     logger: ILogger,
-    cache: ICache<Serializable>,
   ) {
+    this.#tokenHelper = tokenHelper;
     this.#connection = connection;
     this.#transactionHelper = transactionHelper;
     this.#logger = createPrefixedLogger(logger, '[📩 SendSplTokenBuilder]');
-    this.#cache = cache;
   }
 
   async buildTransactionMessage(
@@ -91,40 +68,20 @@ export class SendSplTokenBuilder implements ISendTransactionBuilder {
     const { from, to, mint, amount, network } = params;
     assert(mint, 'Mint is required');
 
-    const rpc = this.#connection.getRpc(network);
+    const mintAccount = await this.#connection.fetchMint(mint, network);
 
-    const fetchMintCached = useCache<
-      [Rpc<GetAccountInfoApi>, Address],
-      Account<Mint, any> & Serializable
-    >(fetchMint as any, this.#cache, {
-      ttlMilliseconds: this.#cacheTtlsMilliseconds.mintAccount,
-      functionName: 'fetchMint',
-    });
-
-    const fetchJsonParsedAccountCached = useCache<
-      [Rpc<GetTokenAccountsByOwnerApi>, Address],
-      MaybeAccount<any> | (MaybeEncodedAccount & Serializable)
-    >(fetchJsonParsedAccount as any, this.#cache, {
-      ttlMilliseconds: this.#cacheTtlsMilliseconds.tokenAccount,
-      functionName: 'fetchJsonParsedAccount',
-    });
-
-    const [mintAccount, splTokenTokenAccount] = await Promise.all([
-      fetchMintCached(rpc, mint),
-      fetchJsonParsedAccountCached(rpc, mint),
-    ]);
-
-    SendSplTokenBuilder.assertAccountExists(splTokenTokenAccount);
-
-    const tokenProgram = splTokenTokenAccount.programAddress;
-    const decimals = this.getDecimals(splTokenTokenAccount);
+    const {
+      programAddress: tokenProgram,
+      data: { decimals },
+    } = mintAccount;
 
     /**
      * The user inputs the amount thinking in uiAmount terms,
      * so if the token uses a multiplier, we need to convert that uiAmount to the raw amount
      */
-    const rawAmountInLamports = TokenHelper.uiAmountToAmountForMint(
-      mintAccount,
+    const rawAmountInLamports = await this.#tokenHelper.uiAmountToAmountForMint(
+      mint,
+      network,
       amount.toString(),
     );
 
@@ -222,88 +179,6 @@ export class SendSplTokenBuilder implements ISendTransactionBuilder {
         tokenProgram,
       })
     )[0];
-  }
-
-  /**
-   * Get the decimals of a given token account.
-   * @param tokenAccount - The token account.
-   * @returns The decimals.
-   */
-  getDecimals<TData extends Uint8Array | MaybeHasDecimals>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): number {
-    SendSplTokenBuilder.assertAccountExists(tokenAccount);
-    SendSplTokenBuilder.assertAccountDecoded(tokenAccount);
-
-    const { decimals } = tokenAccount.data;
-
-    if (!decimals) {
-      throw new Error(`Decimals not found for ${tokenAccount}`);
-    }
-
-    return decimals;
-  }
-
-  /**
-   * Check if a token account exists.
-   * @param tokenAccount - The token account.
-   * @returns Whether the token account exists.
-   */
-  static isAccountExists<TData extends Uint8Array | object>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): boolean {
-    return tokenAccount.exists;
-  }
-
-  /**
-   * Assert that a token account exists.
-   * @param tokenAccount - The token account.
-   */
-  static assertAccountExists<TData extends Uint8Array | object>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): asserts tokenAccount is (MaybeAccount<TData> | MaybeEncodedAccount) &
-    Exists {
-    if (!SendSplTokenBuilder.isAccountExists(tokenAccount)) {
-      throw new Error('Token account does not exist');
-    }
-  }
-
-  /**
-   * Assert that a token account does not exists.
-   * @param tokenAccount - The token account.
-   */
-  static assertAccountNotExists<TData extends Uint8Array | object>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): asserts tokenAccount is (MaybeAccount<TData> | MaybeEncodedAccount) &
-    Exists {
-    if (SendSplTokenBuilder.isAccountExists(tokenAccount)) {
-      throw new Error('Token account exists');
-    }
-  }
-
-  /**
-   * Check if a token account is decoded.
-   * @param tokenAccount - The token account.
-   * @returns Whether the token account is decoded.
-   */
-  static isAccountDecoded<TData extends Uint8Array | object>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): boolean {
-    SendSplTokenBuilder.assertAccountExists(tokenAccount);
-    return !(tokenAccount.data instanceof Uint8Array);
-  }
-
-  /**
-   * Assert that a token account is decoded.
-   * @param tokenAccount - The token account.
-   */
-  static assertAccountDecoded<TData extends Uint8Array | object>(
-    tokenAccount: MaybeAccount<TData> | MaybeEncodedAccount,
-  ): asserts tokenAccount is Account<Exclude<TData, Uint8Array>> & Exists {
-    SendSplTokenBuilder.assertAccountExists(tokenAccount);
-    if (!SendSplTokenBuilder.isAccountDecoded(tokenAccount)) {
-      throw new Error('Token account is encoded. Implement a decoder.');
-    }
   }
 
   getComputeUnitLimit(): number {
